+++ test/cookie_test.exs
@@ -1,12 +1,12 @@
-defmodule ExFirebaseAuth.TokenTest do
+defmodule ExFirebaseAuth.CookieTest do
   use ExUnit.Case
 
   alias ExFirebaseAuth.{
-    Token,
+    Cookie,
     Mock
   }
 
-  defp generate_token(claims, jws) do
+  defp generate_cookie(claims, jws) do
     [{_kid, jwk}] = :ets.lookup(ExFirebaseAuth.Mock, :ets.first(ExFirebaseAuth.Mock))
 
     {_, payload} = JOSE.JWT.sign(jwk, jws, claims) |> JOSE.JWS.compact()
@@ -20,20 +20,20 @@ defmodule ExFirebaseAuth.TokenTest do
 
     on_exit(fn ->
       :ok = Application.delete_env(:ex_firebase_auth, :mock)
-      :ok = Application.delete_env(:ex_firebase_auth, :issuer)
+      :ok = Application.delete_env(:ex_firebase_auth, :cookie_issuer)
     end)
   end
 
-  describe "Token.verify_token/1" do
+  describe "Cookie.verify/1" do
     test "Does succeed on correct token" do
       issuer = Enum.random(?a..?z)
-      Application.put_env(:ex_firebase_auth, :issuer, issuer)
+      Application.put_env(:ex_firebase_auth, :cookie_issuer, issuer)
 
       sub = Enum.random(?a..?z)
       time_in_future = DateTime.utc_now() |> DateTime.add(360, :second) |> DateTime.to_unix()
       claims = %{"exp" => time_in_future}
-      valid_token = Mock.generate_token(sub, claims)
-      assert {:ok, ^sub, jwt} = Token.verify_token(valid_token)
+      valid_token = Mock.generate_cookie(sub, claims)
+      assert {:ok, ^sub, jwt} = Cookie.verify(valid_token)
 
       %JOSE.JWT{
         fields: %{
@@ -47,25 +47,25 @@ defmodule ExFirebaseAuth.TokenTest do
     end
 
     test "Does raise on no issuer being set" do
-      Application.put_env(:ex_firebase_auth, :issuer, "issuer")
-      valid_token = Mock.generate_token("subsub")
-      Application.delete_env(:ex_firebase_auth, :issuer)
+      Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
+      valid_token = Mock.generate_cookie("subsub")
+      Application.delete_env(:ex_firebase_auth, :cookie_issuer)
 
       assert_raise(
         ArgumentError,
-        ~r/^could not fetch application environment :issuer for application :ex_firebase_auth because configuration at :issuer was not set/,
+        ~r/^could not fetch application environment :cookie_issuer for application :ex_firebase_auth because configuration at :cookie_issuer was not set/,
         fn ->
-          Token.verify_token(valid_token)
+          Cookie.verify(valid_token)
         end
       )
     end
 
     test "Does fail on no `kid` being set in JWT header" do
       sub = Enum.random(?a..?z)
-      Application.put_env(:ex_firebase_auth, :issuer, "issuer")
+      Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
 
       token =
-        generate_token(
+        generate_cookie(
           %{
             "sub" => sub,
             "iss" => "issuer"
@@ -75,16 +75,16 @@ defmodule ExFirebaseAuth.TokenTest do
           }
         )
 
-      assert {:error, "Invalid JWT header, `kid` missing"} = Token.verify_token(token)
+      assert {:error, "Invalid JWT header, `kid` missing"} = Cookie.verify(token)
     end
   end
 
   test "Does fail invalid kid being set" do
     sub = Enum.random(?a..?z)
-    Application.put_env(:ex_firebase_auth, :issuer, "issuer")
+    Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
 
     token =
-      generate_token(
+      generate_cookie(
         %{
           "sub" => sub,
           "iss" => "issuer"
@@ -96,12 +96,12 @@ defmodule ExFirebaseAuth.TokenTest do
       )
 
     assert {:error, "Public key retrieved from google was not found or could not be parsed"} =
-             Token.verify_token(token)
+             Cookie.verify(token)
   end
 
   test "Does fail on invalid signature with non-matching kid" do
     sub = Enum.random(?a..?z)
-    Application.put_env(:ex_firebase_auth, :issuer, "issuer")
+    Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
 
     {_invalid_kid, public_key, private_key} = Mock.generate_key()
 
@@ -122,17 +122,17 @@ defmodule ExFirebaseAuth.TokenTest do
       )
       |> JOSE.JWS.compact()
 
-    assert {:error, "Invalid signature"} = Token.verify_token(token)
+    assert {:error, "Invalid signature"} = Cookie.verify(token)
   end
 
   test "Does fail on invalid issuer" do
     sub = Enum.random(?a..?z)
-    Application.put_env(:ex_firebase_auth, :issuer, "issuer")
+    Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
 
     [{kid, _}] = :ets.lookup(ExFirebaseAuth.Mock, :ets.first(ExFirebaseAuth.Mock))
 
     token =
-      generate_token(
+      generate_cookie(
         %{
           "sub" => sub,
           "iss" => "bogusissuer"
@@ -143,28 +143,28 @@ defmodule ExFirebaseAuth.TokenTest do
         }
       )
 
-    assert {:error, "Signed by invalid issuer"} = Token.verify_token(token)
+    assert {:error, "Signed by invalid issuer"} = Cookie.verify(token)
   end
 
   test "Does fail on invalid JWT with raised exception handled" do
-    Application.put_env(:ex_firebase_auth, :issuer, "issuer")
+    Application.put_env(:ex_firebase_auth, :cookie_issuer, "issuer")
 
     invalid_token = "invalid.jwt.token"
 
-    assert {:error, "Invalid JWT"} = Token.verify_token(invalid_token)
+    assert {:error, "Invalid JWT"} = Cookie.verify(invalid_token)
   end
 
   test "Does fail on expired JWT" do
     issuer = Enum.random(?a..?z)
-    Application.put_env(:ex_firebase_auth, :issuer, issuer)
+    Application.put_env(:ex_firebase_auth, :cookie_issuer, issuer)
 
     sub = Enum.random(?a..?z)
 
     time_in_past = DateTime.utc_now() |> DateTime.add(-60, :second) |> DateTime.to_unix()
     claims = %{"exp" => time_in_past}
 
-    valid_token = Mock.generate_token(sub, claims)
+    valid_token = Mock.generate_cookie(sub, claims)
 
-    assert {:error, "Expired JWT"} = Token.verify_token(valid_token)
+    assert {:error, "Expired JWT"} = Cookie.verify(valid_token)
   end
 end